Turn off filemode; manage dirs with .gitkeep
Added .gitkeep and associated rules to .gitignore so that directories
that were empty would be created when I did a git clone on a new
machine. This was prompted by OpenBSD. I also had to change the filemode
on OpenBSD to false. I am not sure what the correct permissions are, but
I didn't want to dirty all those files.