Add xorg.conf to lock TTY and prevent user from killing X
via the Arch wiki:
The /etc/X11/xorg.conf.d/ directory stores host-specific
configuration. You are free to add configuration files there, but
they must have a .conf suffix: the files are read in ASCII order,
and by convention their names start with XX- (two digits and a
hyphen, so that for example 10 is read before 20).